deleted some files
