+++ Makefile
@@ -7,7 +7,7 @@ OBJ_DIR := obj
 BIN_DIR := bin
 
 # List of files to exclude
-EXCLUDE_FILES_SRC := PM3.cpp PM2_full.cpp PM2_one_iteration.cpp rSVD.cpp
+EXCLUDE_FILES_SRC := 
 
 # Find all source files in the source directory excluding specific files
 SRC_FILES := $(filter-out $(addprefix $(SRC_DIR)/,$(EXCLUDE_FILES_SRC)), $(wildcard $(SRC_DIR)/*.cpp))
@@ -15,7 +15,7 @@ SRC_FILES := $(filter-out $(addprefix $(SRC_DIR)/,$(EXCLUDE_FILES_SRC)), $(wildc
 # Create a list of object files based on the source files
 OBJ_FILES := $(patsubst $(SRC_DIR)/%.cpp,$(OBJ_DIR)/%.o,$(SRC_FILES))
 
-EXCLUDE_FILES_TST := PM3Test.cpp PMTest2.cpp SVD_test.cpp PMTest3.cpp rSVD_test.cpp PMTest.cpp
+EXCLUDE_FILES_TST := PM3Test.cpp PMTest2.cpp SVD_test.cpp PMTest3.cpp PMTest.cpp QRTest.cpp
 
 # Test source files
 TEST_SRC_FILES := $(filter-out $(addprefix $(TESTS_DIR)/,$(EXCLUDE_FILES_TST)), $(wildcard $(TESTS_DIR)/*.cpp))
