--- data/output/Q_bcsstm01.mtx_output.mtx
@@ -1,102 +1,0 @@
-%%MatrixMarket matrix coordinate  real general
-10 10 100
-1 1 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 2 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 3 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 4 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 4 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 5 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 5 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 6 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 6 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 7 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 7 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 8 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 8 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 9 1.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 9 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-2 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-4 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-5 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-6 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-7 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-8 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-9 10 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-10 10 1.0000000000000000000000000000000000000000000000000000000000000000e+00
