--- data/output/Q_simple.mtx_output.mtx
@@ -1,11 +1,0 @@
-%%MatrixMarket matrix coordinate  real general
-3 3 9
-1 1 1.2309149097933272387628278465854236856102943420410156250000000000e-01
-2 1 4.9236596391733100652743360114982351660728454589843750000000000000e-01
-3 1 8.6164043685532931693416003327001817524433135986328125000000000000e-01
-1 2 9.0453403373329099856903212639736011624336242675781250000000000000e-01
-2 2 3.0151134457776351815994075877824798226356506347656250000000000000e-01
-3 2 -3.0151134457776357367109199003607500344514846801757812500000000000e-01
-1 3 -4.0824829046386290620773706905310973525047302246093750000000000000e-01
-2 3 8.1649658092772614548238152565318159759044647216796875000000000000e-01
-3 3 -4.0824829046386307274119076282659079879522323608398437500000000000e-01
