--- data/output/R_simple.mtx_output.mtx
@@ -1,11 +1,0 @@
-%%MatrixMarket matrix coordinate  real general
-3 3 9
-1 1 8.1240384046359626069033765816129744052886962890625000000000000000e+00
-2 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-3 1 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 2 9.6011362963879545162626527599059045314788818359375000000000000000e+00
-2 2 9.0453403373329077652442720136605203151702880859375000000000000000e-01
-3 2 0.0000000000000000000000000000000000000000000000000000000000000000e+00
-1 3 1.1078234188139948201978768338449299335479736328125000000000000000e+01
-2 3 1.8090680674665817750934593277634121477603912353515625000000000000e+00
-3 3 0.0000000000000000000000000000000000000000000000000000000000000000e+00
