+++ include/powerMethod/SVD.hpp
@@ -1,7 +1,7 @@
 #ifndef SVD_H
 #define SVD_H
 
-#include "PM1.hpp"
+#include "PM.hpp"
 
 void SVD(Mat &A, Vec &sigma, Mat &U, Mat &V, const int dim);
