+++ include/r_SVD/rSVD.hpp
@@ -6,15 +6,14 @@
 #include <Eigen/Sparse>
 
 
+using Mat = Eigen::MatrixXd;
+using Vet = Eigen::VectorXd;
 
-using mat = Eigen::MatrixXd;
-using vet = Eigen::VectorXd;
-
-using mat=Eigen::MatrixXd;
-using vet=Eigen::VectorXd;
+using Mat=Eigen::MatrixXd;
+using Vet=Eigen::VectorXd;
 
 void intermediate_step(Mat &A,Mat &Q,Mat &Omega,int &l,int &q);
-void rSVD(mat& A, mat&U, vet& S, mat& V);
+void rSVD(Mat& A, Mat&U, Vet& S, Mat& V);
 
 
 #endif
